Move error constant from helper to config
