+++ controllers/setup_test.go
@@ -4,7 +4,7 @@ import (
 	"os"
 	"testing"
 
-	errorHelper "github.com/gutakk/go-google-scraper/helpers/error_handler"
+	errorconf "github.com/gutakk/go-google-scraper/config/error"
 	"github.com/gutakk/go-google-scraper/helpers/log"
 	"github.com/gutakk/go-google-scraper/tests/path_test"
 
@@ -16,7 +16,7 @@ func TestMain(m *testing.M) {
 
 	err := os.Chdir(path_test.GetRoot())
 	if err != nil {
-		log.Fatal(errorHelper.ChangeToRootDirFailure, err)
+		log.Fatal(errorconf.ChangeToRootDirFailure, err)
 	}
 
 	os.Exit(m.Run())
