+++ tests/http/http.go
@@ -7,14 +7,14 @@ import (
 	"net/url"
 	"strings"
 
-	errorHelper "github.com/gutakk/go-google-scraper/helpers/error_handler"
+	errorconf "github.com/gutakk/go-google-scraper/config/error"
 	"github.com/gutakk/go-google-scraper/helpers/log"
 )
 
 func PerformRequest(r http.Handler, method, path string, headers http.Header, payload url.Values) *httptest.ResponseRecorder {
 	req, err := http.NewRequest(method, path, strings.NewReader(payload.Encode()))
 	if err != nil {
-		log.Error(errorHelper.RequestInitializeFailure, err)
+		log.Error(errorconf.RequestInitializeFailure, err)
 	}
 	return perform(req, r, headers)
 }
@@ -22,7 +22,7 @@ func PerformRequest(r http.Handler, method, path string, headers http.Header, pa
 func PerformFileUploadRequest(r http.Handler, method, path string, headers http.Header, payload *bytes.Buffer) *httptest.ResponseRecorder {
 	req, err := http.NewRequest(method, path, payload)
 	if err != nil {
-		log.Error(errorHelper.RequestInitializeFailure, err)
+		log.Error(errorconf.RequestInitializeFailure, err)
 	}
 	return perform(req, r, headers)
 }
